requirement.txt
openai==1.66.5
mysql-connector-python==9.2.0
simplejson==3.20.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
Jinja2==3.1.6
langchain-openai==0.3.9
langchain-community==0.3.20
langchain-upstage==0.6.0
